fix and move unit tests
